feat Allow user to move books between shelves
While the menu was present previously, it was non-functional. This commit allows the user to move books between shelves.
The callbacks update the state of the the App as well as updating the book in the api allowing the change to persist
across page refreshes. I also added a default value to the BookShelfChanger component to ensure the user is able to
move books to anotehr shelf at all times; previously this was implemented such that 'Currently Reading' was default
selected on all shelves.