feat refactor Implement search additions and begin refactor
While the search functionality was present, the user was unable to add books to their shelves. This commit
adds the ability for users to add books to any available shelf. This commit also includes the beginnings
of a refactor. In the previous implementation, there was a lot of unnecessary logic living in App.js to
filter all of the book data from an API call and pass it down. This felt like a poor distribution of responsibility
and I've opted to have the Book object leverage the API directly in order to obtain the data for a book based
on its id. All features are functional at this time.