refactor Condense logic in App.js
Previously there were several functions in App.js to specifically to handle the addition of a new book
versus updating an existing book in the app state including several functions which acted as tests. This
commit condenses the logic to ease readability as well as make future updates more developer-friendly.