refactor Remove unnecessary API call and functions
Since in the case of a book update in state we have access to the book information from previous getAll/search calls,
we already have access to the book information we need. Commit removes unnecessary functions as we're now passing
teh book object as an argument to the update callbacks to accomplish the goal of updating the state of the App
and SearchForm as necessary.